Change longitude latidude long -> double

+++ greenUDB/Models/Garden.cs
@@ -23,9 +23,9 @@ public partial class Garden
 
     public string Description { get; set; } = null!;
 
-    public long Latitude { get; set; }
+    public double Latitude { get; set; }
 
-    public long Longitude { get; set; }
+    public double Longitude { get; set; }
 
     public long Length { get; set; }
